+++ src/components/myPage/Collections/ConfirmModal.jsx
@@ -1,8 +1,8 @@
 import { Button, Flex } from '@mantine/core';
 import { useRecoilValue } from 'recoil';
 import { modals } from '@mantine/modals';
-import { useDeleteUserContentMutation } from '../../hooks/mutations';
-import { userState } from '../../recoil/atom';
+import { useDeleteUserContentMutation } from '../../../hooks/mutations';
+import { userState } from '../../../recoil/atom';
 
 const ConfirmModal = ({ id, listName }) => {
   const email = useRecoilValue(userState);
