+++ src/components/myPage/Collections/index.js
@@ -1,9 +1,10 @@
 export { default as Collections } from './Wrapper';
 
 export { default as Collection } from './Collection';
-export { default as Item } from './Item';
 export { default as ItemTitle } from './ItemTitle';
+export { default as Item } from './Item';
 
+export { default as ConfirmModal } from './ConfirmModal';
 export { default as DateEditor } from './DateEditor';
 export { default as CategoryButton } from './CategoryButton';
 export { default as CategoryButtons } from './CategoryButtons';
