+++ src/components/myPage/index.js
@@ -6,4 +6,3 @@ export { CurrentSubscription } from './CurrentSubscription';
 export { Statistics } from './Statistics';
 
 export { Collections } from './Collections';
-export { default as ConfirmModal } from './ConfirmModal';
